# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "description" : "defaultDeription",
    "version" : "1.0.0-oas3"
  },
  "servers" : [ {
    "url" : "https://0812dce4.ngrok.io"
  } ],
  "paths" : {
...
    "/8API/Screen/Transactions/Approvals/{VoucherType}" : {
      "get" : {
        "tags" : [ "Alerts & Approvals" ],
        "description" : "Approvals",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "VoucherType",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_18"
                },
                "examples" : { }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server      "
          }
        },
        "servers" : [ {
          "url" : "https://3cf66f1e.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://3cf66f1e.ngrok.io"
      } ]
    },
...
    "/8API/Screen/Transactions/AlertApprovalCount" : {
      "get" : {
        "tags" : [ "Alerts & Approvals" ],
        "description" : "AlertApprovalCount",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_19"
                },
                "examples" : { }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server      "
          }
        },
        "servers" : [ {
          "url" : "https://3cf66f1e.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://3cf66f1e.ngrok.io"
      } ]
    },
    "/8API/Screen/Transactions/Approvals" : {
      "get" : {
        "tags" : [ "Alerts & Approvals" ],
        "description" : "Approvals",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_20"
                },
                "examples" : { }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server      "
          }
        },
        "servers" : [ {
          "url" : "https://3cf66f1e.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://3cf66f1e.ngrok.io"
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "inline_response_200_18_data" : {
        "type" : "object",
        "properties" : {
          "Index" : {
            "type" : "integer"
          },
          "Name" : {
            "type" : "string"
          },
          "Purpose" : {
            "type" : "integer"
          },
          "Code" : {
            "type" : "string"
          },
          "CreatedDateTime" : {
            "type" : "string"
          },
          "DateTime" : {
            "type" : "string"
          },
          "ScreenId" : {
            "type" : "integer"
          },
          "Status" : {
            "type" : "integer"
          },
          "AlertId" : {
            "type" : "integer"
          },
          "ModuleId" : {
            "type" : "integer"
          },
          "Snooze" : {
            "type" : "string"
          },
          "RecordId" : {
            "type" : "integer"
          },
          "Message" : {
            "type" : "string"
          },
          "Params" : {
            "type" : "integer"
          }
        },
        "example" : {
          "Status" : 5,
          "Snooze" : "Snooze",
          "Message" : "Message",
          "Params" : 3,
          "Index" : 6,
          "Purpose" : 1,
          "Code" : "Code",
          "DateTime" : "DateTime",
          "Name" : "Name",
          "ModuleId" : 7,
          "AlertId" : 2,
          "ScreenId" : 5,
          "RecordId" : 9,
          "CreatedDateTime" : "CreatedDateTime"
        }
      },
      "inline_response_200_18" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_18_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "Status" : 5,
            "Snooze" : "Snooze",
            "Message" : "Message",
            "Params" : 3,
            "Index" : 6,
            "Purpose" : 1,
            "Code" : "Code",
            "DateTime" : "DateTime",
            "Name" : "Name",
            "ModuleId" : 7,
            "AlertId" : 2,
            "ScreenId" : 5,
            "RecordId" : 9,
            "CreatedDateTime" : "CreatedDateTime"
          }, {
            "Status" : 5,
            "Snooze" : "Snooze",
            "Message" : "Message",
            "Params" : 3,
            "Index" : 6,
            "Purpose" : 1,
            "Code" : "Code",
            "DateTime" : "DateTime",
            "Name" : "Name",
            "ModuleId" : 7,
            "AlertId" : 2,
            "ScreenId" : 5,
            "RecordId" : 9,
            "CreatedDateTime" : "CreatedDateTime"
          } ],
          "message" : "message"
        }
      },
      "inline_response_200_19_data" : {
        "type" : "object",
        "properties" : {
          "AlertsCount" : {
            "type" : "integer"
          },
          "ApprovalsCount" : {
            "type" : "integer"
          }
        },
        "example" : {
          "ApprovalsCount" : 1,
          "AlertsCount" : 6
        }
      },
      "inline_response_200_19" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_19_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "ApprovalsCount" : 1,
            "AlertsCount" : 6
          }, {
            "ApprovalsCount" : 1,
            "AlertsCount" : 6
          } ],
          "message" : "message"
        }
      },
      "inline_response_200_20_data" : {
        "type" : "object",
        "properties" : {
          "ModuleId" : {
            "type" : "integer"
          },
          "ScreenId" : {
            "type" : "integer"
          },
          "AlertPurpose" : {
            "type" : "string"
          },
          "Count" : {
            "type" : "integer"
          },
          "ScreenName" : {
            "type" : "string"
          }
        },
        "example" : {
          "ModuleId" : 6,
          "AlertPurpose" : "AlertPurpose",
          "ScreenId" : 1,
          "ScreenName" : "ScreenName",
          "Count" : 5
        }
      },
      "inline_response_200_20" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_20_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "ModuleId" : 6,
            "AlertPurpose" : "AlertPurpose",
            "ScreenId" : 1,
            "ScreenName" : "ScreenName",
            "Count" : 5
          }, {
            "ModuleId" : 6,
            "AlertPurpose" : "AlertPurpose",
            "ScreenId" : 1,
            "ScreenName" : "ScreenName",
            "Count" : 5
          } ],
          "message" : "message"
        }
      },
...
    }
  }
}